refactor(pidash-ui): extract hooks from App.tsx (687 → 277 lines)
Extract 2 custom hooks from the monolithic App component (CRAP 31,862):

- useMessageHandler: message/event state, handler refs, onMessage switch
  statement (~412 lines) — owns messages, streaming, model, tokens state
- useKeyboardNavigation: keyboard event listener for session switching,
  abort, numbered shortcuts (~69 lines)

App.tsx reduced from 687 to 277 lines. Build succeeds.
Pure mechanical extraction with zero logic/visual changes.

Closes #583
